requirements.txt
# Core dependencies
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2

# Machine Learning
torch==2.0.1
scikit-learn==1.3.0

# API & Web
requests==2.31.0
python-dotenv==1.0.0

# Data processing
tqdm==4.66.1

# Report generation
fpdf==1.7.2
Pillow==10.0.0


